Moved APIError exception to docker.errors module.
